refactor(test-harness): hoist the suite identity cluster to the crate root
SuiteKind, SuiteOutcome, SummarySuite, and ToolchainMatrix are used across the
summary statistics, the benchmark writer, and the summary binary — their lowest
common parent is the crate root. Each becomes a root module; the summary module
keeps only Summary (the render entry, which has no internal users) and the
end-to-end tests.
