+++ solx-benchmark-converter/src/lib.rs
@@ -10,6 +10,10 @@ pub(crate) mod comparison;
 pub(crate) mod input;
 pub(crate) mod output;
 pub(crate) mod role;
+pub(crate) mod suite_kind;
+pub(crate) mod suite_outcome;
+pub(crate) mod summary_suite;
+pub(crate) mod toolchain_matrix;
 pub(crate) mod utils;
 
 pub use crate::benchmark::Benchmark;
@@ -31,8 +35,8 @@ pub use crate::input::testing_time::TestingTimeReport;
 pub use crate::output::Output;
 pub use crate::output::format::Format as OutputFormat;
 pub use crate::output::json::Json as OutputJson;
-pub use crate::output::summary::SuiteKind;
-pub use crate::output::summary::SuiteOutcome;
 pub use crate::output::summary::Summary;
-pub use crate::output::summary::SummarySuite;
-pub use crate::output::summary::ToolchainMatrix;
+pub use crate::suite_kind::SuiteKind;
+pub use crate::suite_outcome::SuiteOutcome;
+pub use crate::summary_suite::SummarySuite;
+pub use crate::toolchain_matrix::ToolchainMatrix;
