+++ solx-benchmark-converter/src/output/summary/mod.rs
@@ -33,150 +33,13 @@ pub mod suite_failures;
 pub mod suite_row;
 pub mod suite_stats;
 pub mod summary_template;
-pub mod toolchain_matrix;
 pub mod top_movers;
 pub mod truncated;
 
-use std::path::PathBuf;
-
-use crate::benchmark::Benchmark;
+use crate::summary_suite::SummarySuite;
 
 use self::suite_stats::SuiteStats;
 use self::summary_template::SummaryTemplate;
-
-pub use self::toolchain_matrix::ToolchainMatrix;
-
-///
-/// How the suite's workflow step ended — the comment must distinguish a
-/// suite that never ran from one that errored, and qualify data written by
-/// a step that then failed. The `Default` exists only so `SuiteStats` can
-/// derive one; `from_suite` always sets the real outcome over it.
-///
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, clap::ValueEnum)]
-pub enum SuiteOutcome {
-    /// The step ran to completion.
-    #[default]
-    Success,
-    /// The step ran but exited nonzero, or was cancelled; any report it wrote
-    /// may be partial.
-    #[value(alias = "cancelled")]
-    Failure,
-    /// The step never ran (an earlier hard failure); not the suite's fault.
-    Skipped,
-}
-
-///
-/// The suites the workflow feeds in — the one place owning each suite's
-/// label, file names, gas gating, and toolchain matrix, so the binary, the
-/// harness writers, and the tests cannot drift apart (a gas-gated Project
-/// suite is unrepresentable).
-///
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SuiteKind {
-    /// solx-tester: deterministic REVM gas, tester naming matrix.
-    Tester,
-    /// Foundry projects: fuzz-noisy gas, project naming matrix.
-    Foundry,
-    /// Hardhat projects: fuzz-noisy gas, project naming matrix.
-    Hardhat,
-}
-
-impl SuiteKind {
-    /// The human-readable suite name shown in the table.
-    pub fn label(self) -> &'static str {
-        match self {
-            Self::Tester => "solx-tester",
-            Self::Foundry => "Foundry",
-            Self::Hardhat => "Hardhat",
-        }
-    }
-
-    /// The XLSX report file name inside the uploaded artifact.
-    pub fn report_file(self) -> &'static str {
-        match self {
-            Self::Tester => "solx-tester-report.xlsx",
-            Self::Foundry => "foundry-report.xlsx",
-            Self::Hardhat => "hardhat-report.xlsx",
-        }
-    }
-
-    /// The merged benchmark JSON file name the summary step reads; the
-    /// workflow spells these paths in `integration-tests.yaml`.
-    pub fn benchmark_file(self) -> &'static str {
-        match self {
-            Self::Tester => "solx-tester-benchmark.json",
-            Self::Foundry => "foundry-benchmark.json",
-            Self::Hardhat => "hardhat-benchmark.json",
-        }
-    }
-
-    /// Whether the suite's gas is deterministic and gates correctness.
-    pub fn gas_is_gate(self) -> bool {
-        matches!(self, Self::Tester)
-    }
-
-    /// Which toolchain naming matrix the suite's mode strings follow.
-    pub fn matrix(self) -> ToolchainMatrix {
-        match self {
-            Self::Tester => ToolchainMatrix::Tester,
-            Self::Foundry | Self::Hardhat => ToolchainMatrix::Project,
-        }
-    }
-
-    ///
-    /// Loads this suite for the summary, or `None` when it was not part of the
-    /// invocation (no benchmark and not skipped). A skipped outcome renders a
-    /// "did not run" row; a benchmark path that will not parse renders the
-    /// suite as errored rather than aborting the summary for the healthy ones.
-    /// A skipped upload step passes its URL through as an empty string, which
-    /// is treated as no URL.
-    ///
-    pub fn load(
-        self,
-        path: Option<PathBuf>,
-        report_url: Option<String>,
-        outcome: SuiteOutcome,
-    ) -> Option<SummarySuite> {
-        let benchmark = match (outcome, path) {
-            (SuiteOutcome::Skipped, _) => None,
-            (_, None) => return None,
-            (_, Some(path)) => match Benchmark::try_from(path.as_path()) {
-                Ok(benchmark) => Some(benchmark),
-                Err(error) => {
-                    eprintln!(
-                        "Warning: {} benchmark is unusable ({error}); rendering the suite as errored.",
-                        self.label()
-                    );
-                    None
-                }
-            },
-        };
-        Some(SummarySuite {
-            kind: self,
-            benchmark,
-            report_url: report_url.filter(|url| !url.is_empty()),
-            outcome,
-        })
-    }
-}
-
-///
-/// One suite (solx-tester / Foundry / Hardhat) fed into the summary.
-///
-pub struct SummarySuite {
-    /// Which suite this is. Its label, report file, gas gating, and toolchain
-    /// matrix all follow from the kind rather than being restated here, so a
-    /// suite cannot contradict its own identity.
-    pub kind: SuiteKind,
-    /// The merged benchmark holding every toolchain's runs. `None` when the
-    /// suite was expected but produced no report (it errored before writing) —
-    /// rendered as an explicit failed row rather than silently dropped.
-    pub benchmark: Option<Benchmark>,
-    /// Artifact download URL for the XLSX report, if uploaded.
-    pub report_url: Option<String>,
-    /// How the suite's workflow step ended.
-    pub outcome: SuiteOutcome,
-}
 
 ///
 /// The suites a single PR summary comment is rendered from.
@@ -205,13 +68,17 @@ impl Summary {
 
 #[cfg(test)]
 mod tests {
+    use crate::benchmark::Benchmark;
     use crate::benchmark::run_failures::RunFailures;
     use crate::benchmark::test::Test;
     use crate::benchmark::test::input::Input as TestInput;
     use crate::benchmark::test::metadata::Metadata;
     use crate::benchmark::test::run::Run;
     use crate::benchmark::test::selector::Selector;
-    use crate::output::summary::*;
+    use crate::output::summary::Summary;
+    use crate::suite_kind::SuiteKind;
+    use crate::suite_outcome::SuiteOutcome;
+    use crate::summary_suite::SummarySuite;
 
     fn render(suites: Vec<SummarySuite>) -> String {
         Summary::new(suites).render()
